rrs-api/src/test/java/kr/co/plateland/rrs/domain/RestaurantTests.java
package kr.co.plateland.rrs.domain;

import org.junit.Test;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.*;

public class RestaurantTests {

    @Test
    public void creation() {
        Restaurant restaurant = new Restaurant(1004L,"Bob Jip", "Seoul");

        assertThat(restaurant.getId(), is(1004L));
        assertThat(restaurant.getName(), is("Bob Jip"));
        assertThat(restaurant.getAddress(), is("Seoul"));
    }

    @Test
    public void information() {
        Restaurant restaurant = new Restaurant(1004L,"Bob Jip", "Seoul");

        assertThat(restaurant.getInformation(), is("Bob Jip in Seoul"));
    }
}